requirements.txt
clause-segmenter==0.1.0
python-docx==1.0.1
reportlab==4.2.2
jsonschema==4.19.2
matplotlib==3.7.1
numpy==1.24.3
openai==0.28
openpyxl==3.1.2
pandas==1.5.3
panel==1.3.1
param==2.0.1
seaborn==0.13.0
tiktoken==0.5.1
jupyter==1.0.0
spacy>=3.0.0,<4.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
